Fix sample with Routing metadata changes
Prior to this commit, changes in the routing metadata mechanism had
broken this sample application.
The routing metadata is now transmitted over the wire and the metadata
must be encoded/decoded according to the MIME type used.

See https://github.com/rsocket/rsocket/blob/master/Extensions/Routing.md

This commit implements a new Serializer that writes the metadata length
and then the metadata itself as an UTF8 string.

This only supports a single type of metadata.
We'll need to use Composite metadata if we want to send other metadata,
such as authentication.