Add airports to the local database
The radar collector is using an in-memory database (for demo purposes)
to store the name, airport code and location of many airports in the
world.

It is using a Spring Data MongoDB reactive repository to store and
search for airports at runtime. This commit also comes with an
initializr that loads data from a JSON file and stores it into the
database.