Add aircraft signal generator
This commit adds `AircraftTrace`, which represents a sequence of
aircraft signals received by a radar. For this, we need various
utilities to calculate the distance, position and bearing in `LatLng`
coordinates.

This is also adding an `AircraftTraceGenerator` that creates random (yet
somehow realistic) data for aircrafts.

Because we need data to not reset itself completely between calls (and
avoid display glitches on the resulting map), the generator caches
generated traces in a Caffeine cache.