Add user profile information
This commit configures the RSocket server using the WebSocket transport
and plugs it into the existing WebFlux server.

It then adds user profile information concepts and stores sample
profiles in the local database.
We're exposing the user profiles as RSocket endpoints. Note that the
current profile is now hardcoded, since we'll need security to figure
out which user is connected to the application.