Add radar endpoints
This commit creates a `RadarService` component that connects to the
radar-collector application and is able to send RSocket requests to it.

We can then expose new RSocket endpoints in this application and fetch
data about specific airports, fetch airports within a particular area
and zip multiple streams of aircrafts signals for several airports.